First iteration for DirectoryViewWidget done. Made a new class ItemGroup to extend WidgetGroup to use layout instead of rebuilding the whole view. Made a new class Item to extend Widget and have custom look, based on the provoded size.
